add dynamic module support
currently, we can only use product which is listed in modules, when
a new product is published or wants to be supported by SDK 2.0, we need
to create a simple file and release a new version for it. Such
workflow is a bit annoying and unwelcome comparing to SDK 3.0, this
patch adds a trick to enable users write code for product which is
not listed in SDK 2.0.